+++ moco_wrapper/exceptions/not_found.py
@@ -1,0 +1,6 @@
+from .base import MocoException
+
+class NotFoundException(MocoException):
+    
+    def __str__(self):
+        return "<NotFoundException, Status Code: {}>".format(self.http_response.status_code)