+++ moco_wrapper/exceptions/server_error.py
@@ -1,0 +1,6 @@
+from .base import MocoException
+
+class ServerErrorException(MocoException):
+        
+    def __str__(self):
+        return "<ServerErrorException, Status Code: {}>".format(self.http_response.status_code)