+++ moco_wrapper/models/objector_models/__init__.py
@@ -22,3 +22,4 @@ from .deal_category import DealCategory
 from .company import Company
 from .comment import Comment
 from .schedule import Schedule
+from .session import Session