+++ moco_wrapper/models/project_expense.py
@@ -198,7 +198,7 @@ class ProjectExpense(MWRAPBase):
                 project_id, 
                 expense_ids_to_disregard, 
                 "Expenses already billed"
-                )
+            )
         """