+++ moco_wrapper/util/objector/__init__.py
@@ -1,2 +1,3 @@
 from .raw import RawObjector
 from .default import DefaultObjector
+from .no_error import NoErrorObjector