+++ moco_wrapper/util/requestor/default.py
@@ -81,6 +81,8 @@ class DefaultRequestor(BaseRequestor):
         elif method == "PATCH":
             response = self.session.patch(path, params=params, json=data, **kwargs)
 
+        print(response)
+
         #convert the reponse into an MWRAPResponse object
         try:
             if response.status_code in self.SUCCESS_STATUS_CODES: